Migrate to Open Telemetry
Open Tracing has now migrated to the Open Telemetry project. In addition to tracing, Open Telemetry also supports metrics. By migrating to Open Telemetry, we can do away with both Open Tracing and DropWizard Metrics and support what looks to be becoming a new standard.

Merging this PR will warrant a semver update to 2.0.0 as we're taking quite a new approach. Open Tracing is relatively easy to migrate (note that "logs" and now "events" on the tracing API). Metrics is a re-write, but a good one given industry support.

JSON serialisation has now been replaced with marshalling to Open Telemetry's standard protobuf domain model. Instead of using `jq` and `nc` to observe services, we'll be able to use `protoc --decode-raw` and `nc` instead.